+++ src/ModelParser/NamingStrategy/PropertyNamingStrategyInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Liip\MetadataParser\ModelParser\NamingStrategy;
+
+interface PropertyNamingStrategyInterface
+{
+    public function getSerializedName(string $name): string;
+}